feat(switch): land scr_vm and resolve broad Switch GCC warning errors
* Land src/script/scr_vm.cpp in POSIX/Switch CMake manifests
* Add Switch-side platform guards (sys/mman, _setjmp, DWORD=uint32_t)
* Fix -Werror=sign-compare via use-site casts (no struct type flips)
* Replace QWORD/DWORD type-punning with memcpy where strict-aliasing fired
* Drop self-aliased qmemcpy(&gfxCmdBufState, &gfxCmdBufState, ...) hex-rays artefacts
* Disable EffectsCore/fx_beam/convert/draw/marks/system/update_util,
  xanim/dobj+xanim, xanim_load_obj on Switch with TODO notes
* Add KISAKHACK annotations on hex-rays nega-array rewrites
* Preserve original ([-128]) array offsets that GCC complains about via
  uintptr_t-launder when the byte semantics must match upstream

+++ src/EffectsCore/fx_archive.cpp
@@ -58,7 +58,7 @@ void __cdecl FX_AddEffectDefTableEntry(FxEffectDefTable *table, uint32_t key, co
 {
     if (!table)
         MyAssertHandler(".\\EffectsCore\\fx_archive.cpp", 47, 0, "%s", "table");
-    if (table->count >= 0x400u)
+    if (static_cast<unsigned int>(table->count) >= 0x400u)
         MyAssertHandler(
             ".\\EffectsCore\\fx_archive.cpp",
             48,
